Define interface `ScheduleNotificationRepository` and its implementation `SharedPreferencesScheduleNotificationRepository`
