+++ app/src/main/java/net/twinte/android/MainApplicationModule.kt
@@ -6,6 +6,8 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import net.twinte.android.repository.schedule.ScheduleRepository
 import net.twinte.android.repository.schedule.SharedPreferencesScheduleRepository
+import net.twinte.android.repository.schedulenotification.ScheduleNotificationRepository
+import net.twinte.android.repository.schedulenotification.SharedPreferencesScheduleNotificationRepository
 import net.twinte.android.repository.user.TwinteBackendUserRepository
 import net.twinte.android.repository.user.UserRepository
 
@@ -21,4 +23,9 @@ interface MainApplicationModule {
     fun bindUserRepository(
         userRepository: TwinteBackendUserRepository,
     ): UserRepository
+
+    @Binds
+    fun bindScheduleNotificationRepository(
+        scheduleNotificationRepository: SharedPreferencesScheduleNotificationRepository,
+    ): ScheduleNotificationRepository
 }
