+++ app/src/main/java/net/twinte/android/repository/schedulenotification/ScheduleNotificationRepository.kt
@@ -1,0 +1,5 @@
+package net.twinte.android.repository.schedulenotification
+
+interface ScheduleNotificationRepository {
+    fun schedule()
+}
