+++ app/src/main/java/net/twinte/android/work/ScheduleNotifier.kt
@@ -1,16 +1,13 @@
 package net.twinte.android.work
 
-import android.app.AlarmManager
 import android.app.PendingIntent
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import android.content.SharedPreferences
 import android.util.Log
 import androidx.core.app.NotificationChannelCompat
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
-import androidx.preference.PreferenceManager
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.runBlocking
 import net.twinte.android.MainActivity
@@ -19,6 +16,7 @@ import net.twinte.android.SettingsActivity
 import net.twinte.android.TWINTE_DEBUG
 import net.twinte.android.model.Day
 import net.twinte.android.repository.schedule.ScheduleRepository
+import net.twinte.android.repository.schedulenotification.ScheduleNotificationRepository
 import java.util.Calendar
 import javax.inject.Inject
 
@@ -29,76 +27,6 @@ import javax.inject.Inject
 class ScheduleNotifier : BroadcastReceiver() {
     @Inject
     lateinit var scheduleRepository: ScheduleRepository
-
-    companion object {
-        private var _preferenceChangeListener: SharedPreferences.OnSharedPreferenceChangeListener? = null
-        private fun getPreferenceChangeListener(context: Context): SharedPreferences.OnSharedPreferenceChangeListener {
-            if (_preferenceChangeListener == null) {
-                _preferenceChangeListener =
-                    SharedPreferences.OnSharedPreferenceChangeListener { _, key ->
-                        if (key != "enable_schedule_notification" && key != "notification_timing") return@OnSharedPreferenceChangeListener
-                        schedule(context)
-                    }
-            }
-            return _preferenceChangeListener!!
-        }
-
-        fun schedule(context: Context) {
-            cancelAll(context)
-            val sharedPreferences = PreferenceManager.getDefaultSharedPreferences(context)
-            sharedPreferences
-                .registerOnSharedPreferenceChangeListener(getPreferenceChangeListener(context))
-            if (!sharedPreferences.getBoolean("enable_schedule_notification", true)) return
-
-            val timing = sharedPreferences.getStringSet("notification_timing", setOf("21", "6"))?.map { it.toInt() }?.sorted()
-                ?: emptyList()
-
-            timing.forEach { hour ->
-                val currentDate = Calendar.getInstance()
-                val notifyDate = Calendar.getInstance().apply {
-                    set(Calendar.HOUR_OF_DAY, hour)
-                    set(Calendar.MINUTE, 0)
-
-                    if (before(currentDate)) {
-                        add(Calendar.HOUR_OF_DAY, 24)
-                    }
-                }
-
-                scheduleAt(context, notifyDate.timeInMillis, hour)
-            }
-        }
-
-        private fun cancelAll(context: Context) {
-            val alarmManager = context.getSystemService(AlarmManager::class.java)
-            for (requestCode in 0..24) {
-                alarmManager.cancel(alarmIntent(context, requestCode))
-            }
-            Log.d("ScheduleNotifier", "All schedule canceled")
-        }
-
-        private fun scheduleAt(context: Context, timeInMillis: Long, requestCode: Int) {
-            val alarmManager = context.getSystemService(AlarmManager::class.java)
-            // 指定された時刻付近で１日おきに設定
-            alarmManager.setRepeating(
-                AlarmManager.RTC_WAKEUP,
-                timeInMillis,
-                1000 * 60 * 60 * 24,
-                alarmIntent(context, requestCode),
-            )
-            Log.d("ScheduleNotifier", "Scheduled at $timeInMillis $requestCode")
-        }
-
-        private fun alarmIntent(context: Context, requestCode: Int) =
-            Intent(context, ScheduleNotifier::class.java).let { intent ->
-                intent.action = "net.twinte.android.action.ScheduleNotifier"
-                PendingIntent.getBroadcast(
-                    context,
-                    requestCode,
-                    intent,
-                    PendingIntent.FLAG_CANCEL_CURRENT or PendingIntent.FLAG_IMMUTABLE,
-                )
-            }
-    }
 
     override fun onReceive(context: Context, intent: Intent?) = runBlocking {
         Log.d("ScheduleNotifier", "Received Broadcast")
@@ -167,10 +95,14 @@ class ScheduleNotifier : BroadcastReceiver() {
         notificationManager.notify(1, notification)
     }
 
-    class OnBootCompleteOrPackageReplaced : BroadcastReceiver() {
+    @AndroidEntryPoint
+    class OnBootCompleteOrPackageReplaced @Inject constructor() : BroadcastReceiver() {
+        @Inject
+        lateinit var scheduleNotificationRepository: ScheduleNotificationRepository
+
         override fun onReceive(context: Context, intent: Intent?) {
             Log.d("ScheduleNotifier", "onReceived ${intent?.action}")
-            schedule(context)
+            scheduleNotificationRepository.schedule()
         }
     }
 }
